fix limit and count bugs

+++ src/include/nodes/pathnodes.h
@@ -1371,8 +1371,8 @@ typedef struct PathKey
 	int			pk_strategy;	/* sort direction (ASC or DESC) */
 	bool		pk_nulls_first; /* do NULLs come before normal values? */
 	
-	int64 limitCount;
-	int64 limitOffset;
+	int limitCount;
+	int limitOffset;
 } PathKey;
 
 /*
